// src/page/admin/optionsUser/UserTypeData.d.ts
export interface UserDataRolName {
	name: string;
}

export interface UserStructure {
	uid: string;
	v_e: string;
	ci: string;
	first_name: string;
	middle_name: string;
	first_surname: string;
	last_surname: string;
	sex: string;
	phone: string;
	email: string;
	uidRol: string;
	rol: UserDataRolName;
}